Return vector<shared_ptr>.
After some tests, it turns out that the compiler is smarter than I assume, without rvo the caller side will receive a new vector obj containing same items as callee side, and all things will stay unchanged with the help of rvo. However the memory of items is in undefined state, which is absolutly unacceptable, so we use shared_ptr instead.